refactor(SerialPrint): ajustar `plot()` para aceptar `int` en lugar de `float` y optimizar cálculos
Se agrega `constexpr pow10()` para manejar potencias de 10. Se comenta la implementación previa de `plot()` con `float` debido a un defecto, y se realiza una conversión a `int` para evitar la manipulación de valores decimales en entornos embebidos. También se actualiza el uso de `millis()` en lugar de `esp_timer_get_time()` para mayor compatibilidad.

+++ lib/SerialPrint/SerialPrint.h
@@ -27,7 +27,7 @@ class SerialPrint {
 public:
     // Datos que queremos graficar
     // template <typename T> static void plot(const char key[], T value, ) { // no recomendable en embedded
-    static void plot(const char key[], float value);
+    static void plot(const char key[], int value);
 
     // Mensajes triviales. Ej.: "¡Enviado exitosamente!"
     static void msg(const char* message);
